test: test and refactor Button component
