+++ src/components/Button/index.jsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const Button = styled.button`
+  width: 100%;
+  min-width: 44px;
+  height: 44px;
+  background-color: var(--secondary);
+  border-radius: 4px;
+  color: var(--primary);
+  font-size: 18px;
+  font-weight: 500;
+  cursor: pointer;
+`;
