+++ src/pages/Details/index.jsx
@@ -1,12 +1,13 @@
 import React, { useContext, useEffect, useState } from "react";
 import { useHistory, useParams } from "react-router-dom";
 import { ButtonWrapper, ButtonContainer } from "./styles";
-import { Container, Button } from "../../styles/GlobalStyles";
+import { Container } from "../../styles/GlobalStyles";
 import { ProductCardDetails } from "../../components/ProductCard";
 import { Store } from "../../context";
 import { Loading } from "../../components/Loading";
 import SearchForm from "../../components/SearchForm";
 import useWindowDimensions from "../../hooks/useWindowDimensions";
+import { Button } from "../../components/Button";
 
 function Details(props) {
   const context = useContext(Store);
